Add basic support for references in attachments, split out msg-content based ones
